+++ src/components/Feed/index.js
@@ -3,3 +3,4 @@ export { FeedBanner } from './FeedBanner/FeedBanner';
 export { SearchBar } from './SearchBar/SearchBar';
 export { HashTags } from './HashTags/HashTags';
 export { PostList } from './PostList/PostList';
+export { FeedAddButton } from './FeedAddButton/FeedAddButton';
