+++ src/pages/FeedListPage/FeedListPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useLocation } from 'react-router-dom';
 import { useInfiniteQuery } from 'react-query';
 import { PageHeader, ScrollContainer } from 'components/Common';
-import { PostList } from 'components/Feed';
+import { FeedAddButton, PostList } from 'components/Feed';
 import { category } from 'api/queries/category';
 import * as S from './styles';
 
@@ -32,6 +32,7 @@ export const FeedListPage = () => {
       <ScrollContainer>
         <PostList infiniteResponse={infiniteResponse} />
       </ScrollContainer>
+      <FeedAddButton />
     </S.PageContainer>
   );
 };
