+++ src/pages/FeedPage/styles.js
@@ -1,4 +1,3 @@
-import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
 export const PageContainer = styled.div`
@@ -26,16 +25,6 @@ export const CategoryWrap = styled.section`
     text-align: center;
     font-size: ${({ theme }) => theme.font.md};
   }
-`;
-
-export const FeedAddBtn = styled(Link)`
-  position: absolute;
-  right: 25px;
-  bottom: 20px;
-  padding: 6px 12px;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-  background-color: ${({ theme }) => theme.color.main};
-  border-radius: 50% 50% 0 50%;
 `;
 
 export const TempLinks = styled.div`
